Promote verify_cookie_len to unsigned before performing arithmetic
verify_cookie_len has type uint8_t and hence is promoted to a
signed int in the arithmetic expression 1 + verify_cookie_len.
This leads to a compiler warning when this value is compared
to the unsigned integer buflen.

+++ library/ssl_cli.c
@@ -1098,7 +1098,7 @@ static int ssl_client_hello_write( mbedtls_ssl_context *ssl,
                               ssl->handshake->verify_cookie,
                               ssl->handshake->verify_cookie_len );
 
-            if( buflen < 1 + ssl->handshake->verify_cookie_len )
+            if( buflen < 1 + (unsigned) ssl->handshake->verify_cookie_len )
             {
                 MBEDTLS_SSL_DEBUG_MSG( 1, ( "buffer too small to hold ClientHello" ) );
                 return( MBEDTLS_ERR_SSL_BUFFER_TOO_SMALL );
